Permit to change the camera settings after compilation
Add CameraConfiguration.h, a class that parse a key=value file and get
the camera parameters. Implemented in a single header in the
SharedHeaders project.

Added static variables to be able to change the camera frame
configuraion. (They are initialized by default to their values in
CameraConstants.h)

Added the loading of the camrea configuration file into the
UnityCompositorInterface. If file isn't found, the old behavior of SV is
not changed.